pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sigil"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "strands-agents",
    "typer[all]",
    "rich",
    "pyyaml",
]

[project.scripts]
sigil = "sigil.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
